+++ src/api/APIConfig.ts
@@ -1,6 +1,11 @@
-import { DataGetterAbstract, DataRefinerAbstract } from "./api.interfaces";
+import {
+  DataGetterAbstract,
+  DataRefinerAbstract,
+  DateService,
+} from "./api.interfaces";
 import APIDataGetter from "./APIDataGetter";
 import APIDataRefiner from "./APIDataRefiner";
+import DateServiceImpl from "./DateServiceImpl";
 import RequestInstance from "./RequestInstance";
 
 export default class APIConfig {
@@ -18,7 +23,11 @@ export default class APIConfig {
     return new APIDataGetter(this.username, this.requireInstance());
   }
 
+  private dateService(): DateService {
+    return new DateServiceImpl();
+  }
+
   public apiDataRefiner(): DataRefinerAbstract {
-    return new APIDataRefiner(this.apiDataGetter());
+    return new APIDataRefiner(this.apiDataGetter(), this.dateService());
   }
 }
